requirements.txt
# LangChain & LangGraph
langchain>=0.1.0
langchain-community>=0.0.10
langchain-anthropic>=0.1.0
langgraph>=0.0.40

# Vector Store
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.12.0

# API & Web
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Data Processing
pandas>=2.1.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.3
